[chore] 친구 추천의 책임 새로운 객체를 만들어 할 (#341)
